+++ src/app/layout.tsx
@@ -1,55 +1,15 @@
 import '@/styles/globals.css'
 import { Inter } from 'next/font/google'
-import { Analytics } from '@vercel/analytics/react'
 
 import * as config from '@/config'
-import { Store } from '@/store/store'
-import '../../sentry.client.config'
-import '../../sentry.server.config'
-import '../../sentry.edge.config'
-import 'react-tooltip/dist/react-tooltip.css'
-import { useInitWeb3InboxClient } from '@web3inbox/widget-react'
 
 const inter = Inter({ subsets: ['latin'] })
 
-var art = `%c
-Слава Україні
- ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⠀⠀⠀⠀⠀⠀
-⠀⠀⠀PEANUT⠀⠀⠀⠀⠀⠀⠀⠀⣀⣤⣶⣶⣦⣌⠙⠋⢡⣴⣶⡄⠀⠀
-⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠀⣿⣿⣿⡿⢋⣠⣶⣶⡌⠻⣿⠟⠀⠀
-⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣿⡆⠸⠟⢁⣴⣿⣿⣿⣿⣿⡦⠉⣴⡇⠀
-⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣾⣿⠟⠀⠰⣿⣿⣿⣿⣿⣿⠟⣠⡄⠹⠀⠀
-⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡀⢸⡿⢋⣤⣿⣄⠙⣿⣿⡿⠟⣡⣾⣿⣿⠀⠀⠀
-⠀⠀⠀⠀⠀⠀⠀⠀⣠⣴⣾⠿⠀⢠⣾⣿⣿⣿⣦⠈⠉⢠⣾⣿⣿⣿⠏⠀⠀⠀
-⠀⠀⠀⠀⣀⣤⣦⣄⠙⠋⣠⣴⣿⣿⣿⣿⠿⠛⢁⣴⣦⡄⠙⠛⠋⠁⠀⠀⠀⠀
-⠀⠀⢀⣾⣿⣿⠟⢁⣴⣦⡈⠻⣿⣿⡿⠁⡀⠚⠛⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
-⠀⠀⠘⣿⠟⢁⣴⣿⣿⣿⣿⣦⡈⠛⢁⣼⡟⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
-⠀⢰⡦⠀⢴⣿⣿⣿⣿⣿⣿⣿⠟⢀⠘⠿⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
-⠀⠘⢀⣶⡀⠻⣿⣿⣿⣿⡿⠋⣠⣿⣷⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
-⠀⠀⢿⣿⣿⣦⡈⠻⣿⠟⢁⣼⣿⣿⠟⠀⠀⠀⠀⠀⠀PEANUT⠀⠀⠀⠀
-⠀⠀⠈⠻⣿⣿⣿⠖⢀⠐⠿⠟⠋⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
-⠀⠀⠀⠀⠈⠉⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀
-`
-
 export default function RootLayout({ children }: { children: React.ReactNode }) {
-    // const web3inboxIsReady = useInitWeb3InboxClient({
-    //     // The project ID and domain you setup in the Domain Setup section
-    //     projectId: process.env.WC_PROJECT_ID ?? '',
-    //     domain: 'peanut.to',
-
-    //     // Allow localhost development with "unlimited" mode.
-    //     // This authorizes this dapp to control notification subscriptions for all domains (including `app.example.com`), not just `window.location.host`
-    //     isLimited: false,
-    // })
     return (
         <html lang="en">
             <body className={inter.className}>
-                <config.ContextProvider>
-                    <Store>
-                        {children}
-                        <Analytics />
-                    </Store>
-                </config.ContextProvider>
+                <config.PeanutProvider>{children}</config.PeanutProvider>
             </body>
         </html>
     )
