--- src/app/utils.ts
@@ -1,9 +1,0 @@
-import { headers } from 'next/headers'
-
-export function currentURL(pathname: string): URL {
-    const headersList = headers()
-    const host = headersList.get('x-forwarded-host') || headersList.get('host')
-    const protocol = headersList.get('x-forwarded-proto') || 'http'
-
-    return new URL(pathname, `${protocol}://${host}`)
-}
