+++ src/config/index.ts
@@ -1,1 +1,2 @@
 export * from './wagmi.config'
+export * from './peanut.config'
